# backend/app/generated/models.py
# generated by fastapi-codegen:
#   filename:  openapi.json
#   timestamp: 2025-01-07T02:31:32+00:00

from __future__ import annotations

from datetime import datetime
from typing import Any, Dict, List, Optional, Union

from pydantic import BaseModel, Field


class BodyUploadDocumentApiDocumentsPost(BaseModel):
    name: str = Field(..., title='Name')
    category: str = Field(..., title='Category')
    description: str = Field(..., title='Description')
    version: Optional[str] = Field('1.0', title='Version')
    file: bytes = Field(..., title='File')


class OpenAIKeyUpdate(BaseModel):
    openai_key: str = Field(..., title='Openai Key')


class ProfileUpdate(BaseModel):
    name: str = Field(..., title='Name')
    jobTitle: Optional[str] = Field(None, title='Jobtitle')
    department: Optional[str] = Field(None, title='Department')


class ScanRequest(BaseModel):
    project_name: str = Field(..., title='Project Name')
    trivy_data: Dict[str, Any] = Field(..., title='Trivy Data')
    openai_key: str = Field(..., title='Openai Key')


class TokenRequest(BaseModel):
    name: str = Field(..., title='Name')
    description: str = Field(..., title='Description')
    expires_in_days: Optional[int] = Field(30, title='Expires In Days')


class TokenResponse(BaseModel):
    token: str = Field(..., title='Token')
    expires_at: datetime = Field(..., title='Expires At')
    id: int = Field(..., title='Id')
    name: str = Field(..., title='Name')


class ValidationError(BaseModel):
    loc: List[Union[str, int]] = Field(..., title='Location')
    msg: str = Field(..., title='Message')
    type: str = Field(..., title='Error Type')


class HTTPValidationError(BaseModel):
    detail: Optional[List[ValidationError]] = Field(None, title='Detail')